# Replace None with an object NoSharding.