This is to make the change to a C++ ShardingSpec easier.