@gnecula 2021-01-18
This LGTM, I asked @skye to take a look also. 

I am assuming that `pxla.ShardingSpec` is not considered public API. Do users ever need to construct or read those objects?