@jblespiau 2021-01-18
It's also internal for me, but there are  few calls from e.g. tensorflow_models, ml_perf. It seems the be the same snippet of code which has been forked with a train script (there are like ~18 such non-jax files doing so, but it does not seem they use None):

```
def _unbroadcast(x):
  """Assuming `x` is replicated along its leading axis, remove that axis."""
  assert isinstance(x, jax.pxla.ShardedDeviceArray)
  sharding_spec = x.sharding_spec
  assert sharding_spec.shards_per_axis[0] == x.shape[0]
  assert not sharding_spec.is_axis_materialized[0]
  aval = jax.abstract_arrays.ShapedArray(x.shape[1:], x.dtype)
  sharding_spec = jax.pxla.ShardingSpec(
      sharding_spec.shards_per_axis[1:],
      sharding_spec.is_axis_materialized[1:],
      [(x.shape[0], 0)] + [(factor, index - 1) for factor, index
                           in sharding_spec.replication_factors])
  return jax.pxla.ShardedDeviceArray(aval, sharding_spec, x.device_buffers)
```

I launched a GTP in cr/352116408 to check nothing fails, assuming tests exist.