@jblespiau 2021-01-20
I ran it internally, because it would take me a lot of time to do it in OS (I would need to build jax locally, which I don't know how to do).

I think it does not change anything, just the usual noise.

Without the change:

```
└──╼ TARGET_TOTAL_SECS=2 CUDA_VISIBLE_DEVICES= XLA_FLAGS=--xla_force_host_platform_device_count=500 bb pmap_benchmark
...

---------Benchmark summary for pmap_shard_sharded_device_array---------
  nargs    nshards      mean       %std    relative
-------  ---------  --------  ---------  ----------
     10          8  0.105513  11.4454       1
    100          8  0.368193   8.90511      3.48956
    101          8  0.356172   6.58713      3.37563
    500          8  1.14319    0.237879    10.8346
   1000          8  2.04005    0           19.3347
   5000          8  8.01584    0           75.9704
    100          2  0.241554   6.35878      2.28934
    100          4  0.294926   6.34386      2.79517
    100          8  0.369116   6.62533      3.49831
    100        100  1.71755    1.24343     16.2782
    100        500  6.70799    0           63.5753

---------Benchmark summary for pmap_shard_device_array---------
  nargs    nshards       mean      %std    relative
-------  ---------  ---------  --------  ----------
     10          8  0.0590367  18.0683      1
    100          8  0.246328    8.73276     4.17245
    500          8  1.05955     1.05783    17.9473
    100          2  0.101085   21.8612      1.71224
    100          4  0.174109   10.2306      2.94915
    100          8  0.250661   11.4526      4.24584

---------Benchmark summary for pmap_shard_outputs---------
  nouts    nshards       mean      %std    relative
-------  ---------  ---------  --------  ----------
     10          8   0.137291  10.5298      1
    100          8   0.413141   7.95815     3.00923
    500          8   1.41002    1.23046    10.2703
   1000          8   2.6428     0          19.2496
   5000          8  11.6036     0          84.518
    100          2   0.266749   9.34605     1.94294
    100          4   0.3165     8.00129     2.30531
    100          8   0.404103   6.20919     2.94339
    100        100   2.41844    0          17.6154
    100        500  12.2105     0          88.9383

---------Benchmark summary for ShardedDeviceArray_indexing---------
indices_fn               mean     %std    relative
------------------  ---------  -------  ----------
integer_indices     0.0719787  12.1457       1
integer_2D_indices  9.12722     0          126.804
```

With the change:

```
└──╼ TARGET_TOTAL_SECS=2 CUDA_VISIBLE_DEVICES= XLA_FLAGS=--xla_force_host_platform_device_count=500 bb pmap_benchmark
---------Benchmark summary for pmap_shard_sharded_device_array---------
  nargs    nshards       mean       %std    relative
-------  ---------  ---------  ---------  ----------
     10          8  0.0996357  13.1614       1
    100          8  0.351055   10.8804       3.52339
    101          8  0.366211    4.27462      3.6755
    500          8  1.09212     2.30705     10.9611
   1000          8  1.94619     1.09397     19.5331
   5000          8  7.59713     0           76.2491
    100          2  0.219362   12.8559       2.20164
    100          4  0.279566    6.64543      2.80589
    100          8  0.364913    5.00456      3.66248
    100        100  1.66384     0.624914    16.6992
    100        500  6.5926      0           66.1671

---------Benchmark summary for pmap_shard_device_array---------
  nargs    nshards       mean       %std    relative
-------  ---------  ---------  ---------  ----------
     10          8  0.0567073  23.0601       1
    100          8  0.226417    7.28242      3.99274
    500          8  1.03926     0.354028    18.3267
    100          2  0.0899203  19.5304       1.58569
    100          4  0.163566   17.4084       2.88438
    100          8  0.247211    6.71917      4.35942

---------Benchmark summary for pmap_shard_outputs---------
  nouts    nshards       mean      %std    relative
-------  ---------  ---------  --------  ----------
     10          8   0.12988   9.07826      1
    100          8   0.408531  9.56531      3.14544
    500          8   1.36306   0.714324    10.4947
   1000          8   2.62184   0           20.1866
   5000          8  11.4392    0           88.0747
    100          2   0.262397  7.61874      2.02029
    100          4   0.30444   9.01886      2.344
    100          8   0.388523  6.51703      2.99139
    100        100   2.42062   0           18.6373
    100        500  11.9683    0           92.1488

---------Benchmark summary for ShardedDeviceArray_indexing---------
indices_fn               mean     %std    relative
------------------  ---------  -------  ----------
integer_indices     0.0657426  15.4291       1
integer_2D_indices  8.35978     0          127.159

```